refactor: split the code of EditProfileModal to multiple files
KK-1071
